+++ CHANGELOG.md
@@ -9,6 +9,7 @@ file. This change log follows the conventions of
 ### Changed
 
 - the `login` function on all handlers now takes a `&mut self` instead of `self`.
+- the `default_states` property of `PlugDeviceInfoResult` is now a struct instead of an enum.
 
 ## [v0.7.2] - 2023-08-21